+++ src-tauri/Cargo.toml
@@ -15,15 +15,7 @@ tauri-build = { git = "https://github.com/tauri-apps/tauri", branch = "next" }
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = [ "derive" ] }
-tauri = { git = "https://github.com/tauri-apps/tauri", rev = "84f6e3e84a34b01b7fa04f5c4719acb921ef4263", features = [
-  "dialog-open",
-  "fs-create-dir",
-  "fs-read-dir",
-  "fs-read-text-file",
-  "fs-write-file",
-  "path-all",
-  "shell-all"
-], version = "1.0.0-beta.8" }
+tauri = { git = "https://github.com/tauri-apps/tauri", rev = "84f6e3e84a34b01b7fa04f5c4719acb921ef4263", features = ["dialog-open", "fs-create-dir", "fs-read-dir", "fs-read-text-file", "fs-write-file", "path-all", "shell-all", "updater"], version = "1.0.0-beta.8" }
 
 [features]
 default = [ "custom-protocol" ]
