+++ src/config/editor-options.js
@@ -1,0 +1,27 @@
+const editorOptions = {
+    lineHeight: 32,
+    fontSize: 14,
+    // fontLigatures: true,
+    // fontFamily: 'Fira Code',
+    contextmenu: false,
+    smoothScrolling: true,
+    scrollBeyondLastLine: true,
+    wordWrap: 'on',
+    wrappingStrategy: 'advance',
+    automaticLayout: true,
+    renderLineHighlight: 'none',
+    renderWhitespace: false,
+    scrollbar: {
+        verticalScrollbarSize: 0,
+        verticalSliderSize: 0
+    },
+    semanticHighlighting: {
+        enabled: true
+    },
+    minimap: {
+        enabled: false
+    },
+    lineNumbersWidth: 0,
+}
+ 
+export default editorOptions;
